backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.16.2
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.5.0
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
gunicorn==22.0.0
werkzeug==3.0.3
